// src/app/models/config.model.ts
export class Config {
  /*   service_url: string = 'http://192.168.1.52/eosce_laravel2';
    name: string = 'Open eOSCE dev';
    client_id: string = '3';
    client_secret: string = 'AhQjIqv7TzuYoTNTFe36tZNPaQLEfHKbzeBLhyEK'; */
    service_url: string;
    name: string;
    client_id: string;
    client_secret: string;
  }